LibAudio: Factorize stream initialisation to base class `LoaderPlugin`
All actual plugins follow the same logic to initialize their stream,
this commit factorizes all of this to their base class: `LoaderPlugin`.
